Big tidy up of Css, moved where the fetch request was happening
