+++ src/App.jsx
@@ -1,25 +1,11 @@
 import './App.css'
 import Navbar from "./Components/Navbar/index.jsx";
 import {useEffect, useState} from "react";
-import ProductsContainer from "./Components/Products Container/index.jsx";
+import ProductsContainer from "./Components/Product/index.jsx";
+import Product from "./Components/Product/index.jsx";
 
-const url = 'https://fakestoreapi.com/products';
 function App() {
 
-const[products, setProducts] = useState([])
-
- const fetchProducts = async () => {
-    try{
-     const response = await fetch(url)
-     const products = await response.json()
-        setProducts(products)
-    } catch(err) {
-        console.log(err)}
- }
-
-    useEffect(() => {
-        fetchProducts()
-    }, []);
   return (
     <>
         <Navbar />
@@ -27,7 +13,7 @@ const[products, setProducts] = useState([])
             <p>Welcome to Cosh and Josmin. yeah!</p>
         </div>
         <main className='main'>
-            <ProductsContainer products={products}/>
+            <Product />
         </main>
     </>
   )
