+++ src/Components/Navbar/index.jsx
@@ -8,7 +8,7 @@ const NavBar = () => {
     return (
         <div>
             <nav className="navBar">
-                <h3>Cosh And Josmin</h3>
+                <p className="welcomeLogo">Cosh And Josmin</p>
                 <CategoryMenu text="Categories" menuItems={['Item 1', 'Item 2', 'Item 3', 'Item 4']} />
                 <SearchBar/>
                 <CartButton text="Add to cart" />
