+++ src/Components/Product/Product.css
@@ -1,0 +1,28 @@
+.products {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    align-items: stretch;
+    justify-content: space-around;
+    margin: 20px;
+    gap:20px;
+}
+
+.single-product {
+    flex: 1 0 calc(40% - 20px); /* Adjusted width and accounting for the gap */
+    background-color: #bbb2b2;
+    padding: 5px;
+    gap: 5px;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+
+}
+
+.single-product .img {
+    background-color: #ffffff;
+    max-height: 150px; /* Set a maximum height for the image */
+    object-fit: contain;
+    border: 1px solid rgba(16, 15, 15, 0.5);
+    border-radius: var(--borderRadius);
+}
