+++ src/Components/Product/index.jsx
@@ -1,13 +1,33 @@
+import ProductContainer from "../ProductContainer/index.jsx";
+import {useEffect, useState} from "react";
+import './Product.css'
+const Product = () => {
+    const url = 'https://fakestoreapi.com/products';
+    const[products, setProducts] = useState([])
 
-const Product = ({id, category, title, price, image, description}) => {
+    const fetchProducts = async () => {
+        try{
+            const response = await fetch(url)
+            const products = await response.json()
+            setProducts(products)
+        } catch(err) {
+            console.log(err)}
+    }
+
+    useEffect(() => {
+        fetchProducts()
+    }, []);
     return (
-        <article className='single-product'>
-        <img src={image} alt={image} className='img'/>
-         <span className='product-price'>£{price}</span>
-         <h2>{title}</h2>
-         <h5>{category}</h5>
-         <p>{description}</p>
-        </article>
-    )
-}
-export default Product
+        <section>
+            <div className='products'>
+                {products.slice(0, 4).map((product) => (
+                    <div className="single-product" key={product.id}>
+                        <ProductContainer {...product} />
+                    </div>
+                ))}
+            </div>
+        </section>
+    );
+};
+
+export default Product;
