+++ src/index.css
@@ -237,23 +237,22 @@ ul {
   margin: 0 auto;
   margin-top: 1rem;
 }
-.single-product{
-  background: var(--white);
-  border-radius: var(--borderRadius);
-  box-shadow: var(--shadow-1);
-  transition: var(--transition);
-  position: relative;
-}
-.single-product .img {
-  height: 20rem;
-  border-top-right-radius: var(--borderRadius);
-  border-top-left-radius: var(--borderRadius);
-}
-.products {
-  padding: 2rem 0;
-  display: flex;
-  gap: 1rem;
-  flex-wrap: wrap;
-  flex-direction:column;
+/*.single-product{*/
+/*  background: var(--white);*/
+/*  border-radius: var(--borderRadius);*/
+/*  box-shadow: var(--shadow-1);*/
+/*  transition: var(--transition);*/
+/*}*/
+/*.single-product .img {*/
+/*  border-top-right-radius: var(--borderRadius);*/
+/*  border-top-left-radius: var(--borderRadius);*/
+/*}*/
+/*.products {*/
+/*  width: 50%;*/
+/*  padding: 2rem 0;*/
+/*  display: flex;*/
+/*  gap: 1rem;*/
+/*  flex-wrap: wrap;*/
+/*  flex-direction:row;*/
+/*}*/
 
-}
